WIP - This is working.
There was a compile time message that System.Runtime was needed for the
CalculationTests class that the System.Runtime dll was needed. In all honesty
I was quite lost for some time about where/what that was and where it came
from. There were several red herrings that were very confusing.

In the end, it's a package in nuget that needs to be pulled in, named
System.Runtime. I swear I had looked for it before, but alas... there it is.