WIP: Initial set of code for EncapsulateClassesWithFactory refactoring
Encapsulate Classes With Factory is found on p.80 of Refatoring To Patterns
This is the initial set of code with tests excercising the code.

You'll notice that the tests use a testing subclass to gain access to
the AttributeDescriptors. A good review of usages of testing subclasses
can be found at:

http://xunitpatterns.com/Test-Specific%20Subclass.html