Duration test is in.
I didn't use his value, as I was getting bogged down in pieces
of the calculation that I don't have I don't think, like how
commitment affects it for example.

I'm a little rushed and so I wanted
to make sure to get a test in for that code path, as that's
the most important thing for now as we work through this Strategy
refactoring.

Given I have the time, I will try to circle back around and
fidget with it.