Introduce an AppendContentsTo as a seam to enable the recursion needed
ToString can't take a StringBuffer because it's an overload and we don't
control the contract for ToString.

Creating an AppendContentsTo Composit method that does all of the work
for us gives us the control we need and allows us to pass in a StringBuffer
class as a parameter. The result that is modified in AppendContentsTo
is then returned out of ToString.

The Accumulating Parameter, result, accumulates all of the mutations needed
along the way as we continue to extract methods now.